Move ContextMenuProvider into its own file (#2055)
